+++ CHANGELOG.md
@@ -5,8 +5,13 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 ## \[Unreleased\]
 
 ### Added
+- TryCP: Calls for clone cell management
+- Tests of clone cell management for local and TryCP conductors
 ### Removed
+- TryCP: Deprecated Admin API request to create clone cells
 ### Changed
+- Upgrade to Holochain JS client v0.9.2 with clone cell features
+- Upgrade to Holochain v0.0.165
 ### Fixed
 
 ## 2022-09-12 [0.8.0]